+++ packages/opentelemetry-instrumentation-reactor/CHANGELOG.md
@@ -1,3 +1,23 @@
+## 6.2.0-rc.8 (2026-07-08)
+
+This was a version bump only for @powerhousedao/opentelemetry-instrumentation-reactor to align it with other projects, there were no code changes.
+
+## 6.2.0-rc.7 (2026-07-07)
+
+This was a version bump only for @powerhousedao/opentelemetry-instrumentation-reactor to align it with other projects, there were no code changes.
+
+## 6.2.0-rc.6 (2026-07-03)
+
+This was a version bump only for @powerhousedao/opentelemetry-instrumentation-reactor to align it with other projects, there were no code changes.
+
+## 6.2.0-rc.5 (2026-07-02)
+
+This was a version bump only for @powerhousedao/opentelemetry-instrumentation-reactor to align it with other projects, there were no code changes.
+
+## 6.2.0-rc.4 (2026-07-01)
+
+This was a version bump only for @powerhousedao/opentelemetry-instrumentation-reactor to align it with other projects, there were no code changes.
+
 ## 6.2.0-rc.3 (2026-06-30)
 
 ### 🚀 Features